+++ hermes-five/src/devices/input/digital.rs
@@ -21,7 +21,7 @@ pub struct DigitalInput {
     // ########################################
     // # Basics
     /// The pin (id) of the [`Board`] used to read the digital value.
-    pin: u16,
+    pin: u8,
     /// The current digital state.
     #[cfg_attr(feature = "serde", serde(with = "crate::devices::arc_rwlock_serde"))]
     state: Arc<RwLock<bool>>,
@@ -71,7 +71,7 @@ impl DigitalInput {
     // Getters and Setters
 
     /// Returns the pin (id) used by the device.
-    pub fn get_pin(&self) -> u16 {
+    pub fn get_pin(&self) -> u8 {
         self.pin
     }
