+++ hermes-five/src/devices/output/digital.rs
@@ -20,7 +20,7 @@ pub struct DigitalOutput {
     // ########################################
     // # Basics
     /// The pin (id) of the [`Board`] used to control the output value.
-    pin: u16,
+    pin: u8,
     /// The current output state.
     #[cfg_attr(feature = "serde", serde(with = "crate::devices::arc_rwlock_serde"))]
     state: Arc<RwLock<bool>>,
@@ -88,7 +88,7 @@ impl DigitalOutput {
     // Setters and Getters.
 
     /// Returns the pin (id) used by the device.
-    pub fn get_pin(&self) -> u16 {
+    pub fn get_pin(&self) -> u8 {
         self.pin
     }
