+++ hermes-five/src/mocks/mod.rs
@@ -11,7 +11,7 @@ pub mod plugin_io;
 pub mod serial_port;
 pub mod transport_layer;
 
-pub fn create_analog_pin(id: u16, value: u16) -> Pin {
+pub fn create_analog_pin(id: u8, value: u16) -> Pin {
     Pin {
         id,
         name: format!("A{}", id),
@@ -42,7 +42,7 @@ pub fn create_analog_pin(id: u16, value: u16) -> Pin {
     }
 }
 
-pub fn create_digital_pin(id: u16, value: u16) -> Pin {
+pub fn create_digital_pin(id: u8, value: u16) -> Pin {
     Pin {
         id,
         name: format!("D{}", id),
@@ -73,7 +73,7 @@ pub fn create_digital_pin(id: u16, value: u16) -> Pin {
     }
 }
 
-pub fn create_input_pin(id: u16, value: u16) -> Pin {
+pub fn create_input_pin(id: u8, value: u16) -> Pin {
     Pin {
         id,
         name: format!("D{}", id),
@@ -100,7 +100,7 @@ pub fn create_input_pin(id: u16, value: u16) -> Pin {
     }
 }
 
-pub fn create_pwm_pin(id: u16, value: u16) -> Pin {
+pub fn create_pwm_pin(id: u8, value: u16) -> Pin {
     Pin {
         id,
         name: format!("D{}", id),
@@ -131,7 +131,7 @@ pub fn create_pwm_pin(id: u16, value: u16) -> Pin {
     }
 }
 
-pub fn create_shift_pin(id: u16, value: u16) -> Pin {
+pub fn create_shift_pin(id: u8, value: u16) -> Pin {
     Pin {
         id,
         name: format!("D{}", id),
@@ -158,7 +158,7 @@ pub fn create_shift_pin(id: u16, value: u16) -> Pin {
     }
 }
 
-pub fn create_servo_pin(id: u16, value: u16) -> Pin {
+pub fn create_servo_pin(id: u8, value: u16) -> Pin {
     Pin {
         id,
         name: format!("D{}", id),
@@ -185,7 +185,7 @@ pub fn create_servo_pin(id: u16, value: u16) -> Pin {
     }
 }
 
-pub fn create_unsupported_pin(id: u16) -> Pin {
+pub fn create_unsupported_pin(id: u8) -> Pin {
     Pin {
         id,
         name: format!("A{}", id),
